+++ mex/editor/search/main.py
@@ -1,22 +1,21 @@
+from typing import Any
+
 import reflex as rx
 
 from mex.common.types import IDENTIFIER_PATTERN
 from mex.editor.component_option_helper import (
-    build_pagination_options,
+    build_pagination_for_state_options,
 )
 from mex.editor.layout import page
-from mex.editor.search.models import (
-    ReferenceFieldIdentifierFilter,
-    SearchPrimarySource,
-)
+from mex.editor.search.models import ReferenceFieldIdentifierFilter, SearchPrimarySource
 from mex.editor.search.state import SearchState, full_refresh
-from mex.editor.search.value_label_select import value_label_select
 from mex.editor.search_results_component import (
     SearchResultsComponentOptions,
     SearchResultsListItemOptions,
     SearchResultsListOptions,
     search_results_component,
 )
+from mex.editor.value_label_select import value_label_select
 
 
 def search_input() -> rx.Component:
@@ -59,16 +58,17 @@ def search_input() -> rx.Component:
     )
 
 
-def entity_type_choice(choice: tuple[str, bool]) -> rx.Component:
+def entity_type_choice(choice: dict[str, Any]) -> rx.Component:
     """Render a single checkbox for filtering by entity type."""
     return rx.checkbox(
-        choice[0],
-        checked=choice[1],
+        choice["label"],
+        checked=choice["checked"],
         on_change=[
-            SearchState.set_entity_type(choice[0]),  # type: ignore[operator]
+            SearchState.set_entity_type(choice["value"]),  # type: ignore[operator]
             *full_refresh,
         ],
         disabled=SearchState.is_loading,
+        custom_attrs={"data-testid": f"entity-type-{choice['value']}"},
     )
 
 
@@ -84,7 +84,7 @@ def entity_type_filter() -> rx.Component:
         ),
         rx.vstack(
             rx.foreach(
-                SearchState.entity_types,
+                SearchState.label_entity_types,
                 entity_type_choice,
             ),
             custom_attrs={"data-testid": "entity-types"},
@@ -103,6 +103,7 @@ def primary_source_choice(choice: tuple[str, SearchPrimarySource]) -> rx.Compone
             *full_refresh,
         ],
         disabled=SearchState.is_loading,
+        custom_attrs={"data-testid": f"primary-source-filter-{choice[0]}"},
     )
 
 
@@ -113,7 +114,7 @@ def primary_source_filter() -> rx.Component:
             SearchState.had_primary_sources,
             primary_source_choice,
         ),
-        custom_attrs={"data-testid": "had-primary-sources"},
+        custom_attrs={"data-testid": "primary-source-filter"},
         style=rx.Style(width="100%"),
     )
 
@@ -286,12 +287,15 @@ def search_results() -> rx.Component:
                 list_options=SearchResultsListOptions(
                     item_options=SearchResultsListItemOptions(enable_title_href=True)
                 ),
-                pagination_options=build_pagination_options(
+                pagination_options=build_pagination_for_state_options(
                     SearchState,
                     SearchState.push_search_params,  # type: ignore[arg-type]
                 ),
             ),
-            style=rx.Style(flex=1),
+            style=rx.Style(
+                flex=1,
+                width="75%",
+            ),
         ),
     )