+++ mex/mex.py
@@ -1,8 +1,8 @@
 import reflex as rx
 from reflex.components.radix import themes
-from reflex.utils.console import info as log_info
 
-from mex.common.logging import logger
+from mex.editor.advanced_search.main import index as advanced_search_index
+from mex.editor.advanced_search.state import AdvancedSearchState
 from mex.editor.api.main import api as editor_api
 from mex.editor.consent.main import index as consent_index
 from mex.editor.consent.state import ConsentState
@@ -24,7 +24,14 @@ from mex.editor.utils import load_settings
 
 app = rx.App(
     theme=themes.theme(accent_color="blue", has_background=False),
-    style={">a": {"opacity": "0"}},
+    style={
+        ">a": {"opacity": "0"},
+        ".truncate": {
+            "overflow": "hidden",
+            "text-overflow": "ellipsis",
+            "white-space": "nowrap",
+        },
+    },
     api_transformer=editor_api,
 )
 app.add_page(
@@ -41,6 +48,17 @@ app.add_page(
     ],
 )
 app.add_page(
+    advanced_search_index,
+    route="/advanced-search",
+    title="MEx Editor | Advanced Search",
+    on_load=[
+        State.check_mex_login,
+        State.load_nav,
+        AdvancedSearchState.search,
+        AdvancedSearchState.resolve_identifiers,
+    ],
+)
+app.add_page(
     merge_index,
     route="/merge",
     title="MEx Editor | Merge",
@@ -53,7 +71,7 @@ app.add_page(
 )
 app.add_page(
     create_index,
-    route="/create/[draft_identifier]",
+    route="/create/[draft_id]",
     title="MEx Editor | Create",
     on_load=[
         State.check_mex_login,
@@ -76,7 +94,7 @@ app.add_page(
 )
 app.add_page(
     edit_index,
-    route="/item/[identifier]",
+    route="/item/[item_id]",
     title="MEx Editor | Edit",
     on_load=[
         State.check_mex_login,
@@ -99,8 +117,16 @@ app.add_page(
         IngestState.resolve_primary_source_titles,
     ],
 )
-app.add_page(login_mex_index, route="/login", title="MEx Editor | Login")
-app.add_page(login_ldap_index, route="/login-ldap", title="MEx Editor | Login")
+app.add_page(
+    login_mex_index,
+    route="/login",
+    title="MEx Editor | Login",
+)
+app.add_page(
+    login_ldap_index,
+    route="/login-ldap",
+    title="MEx Editor | Login",
+)
 app.add_page(
     consent_index,
     route="/consent",
@@ -113,9 +139,5 @@ app.add_page(
     ],
 )
 app.register_lifespan_task(
-    lambda: logger.info(load_settings().text()),
-)
-app.register_lifespan_task(
-    log_info,
-    msg="MEx Editor is running, shut it down using CTRL+C",
+    load_settings,
 )