+++ CHANGELOG.md
@@ -9,13 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Warn about unsaved changes when logging out
-- pytest-retry since tests often failing on github (we belief due to low system/cpu performance in CI)
-
 ### Changes
 
+<<<<<<< feature/mx-2093-use-primary-source-title-on-ingest-tab
 - use dynamic primary source title for ingest tabs
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/81bc5b
+=======
+- change the docker build to multi-stage build
+- change primary source of rules from mex to mex-editor
+- moved pytest rerun config from toml to makefile
+>>>>>>> main
 
 ### Deprecated
 
@@ -23,7 +26,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fixed settings logging on startup and in tests
+
 ### Security
+
+## [1.5.1] - 2026-02-25
+
+### Fixed
+
+- fixed width of search and advanced-search result lists
+- fixed redirect target after login screen interception
+
+## [1.5.0] - 2026-02-24
+
+### Added
+
+- New page for advanced searching, including complex reference filter
+- Translations for entity types
+- Button on edit page to reset or delete rules
+- Warn about unsaved changes when logging out
+- pytest-retry since tests often failing on github (we belief due to low system/cpu performance in CI)
+
+### Changes
+
+- upgrade reflex to 0.8.15
+- use pydantic v2 BaseModel instead of rx.Base
+- other rerun lib for pytest
+- updated mex-common to 1.16.0
+- updated template to https://github.com/robert-koch-institut/mex-template/commit/81bc5b
+
+### Removed
+
+- all custom scripts due to reflex docs
 
 ## [1.4.1] - 2026-02-04